requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
celery>=5.3.0
django-celery-results>=2.5.0
pandas>=2.0.0
redis>=4.6.0 